+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Billingoptions to manifest
+- Add Questions and Answers Bazaarvoice component
 
 ## [1.9.0] - 2021-01-15