+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Billingoptions to manifest
+- Add Questions and Answers Bazaarvoice component
+
 ## [1.9.0] - 2021-01-15
 
 ### Added